Refactor Model Page with Header + Routing
