+++ frontend/src/components/layout/Header.tsx
@@ -1,0 +1,6 @@
+export default function Header() {
+    return (
+        <>
+        </>
+    )
+}
